requirements.txt
# Core
python-dotenv==1.2.2
requests==2.33.1

# Email
resend==2.29.0

# SMS
africastalking==1.2.8

# CRM
hubspot-api-client==12.0.0

# Calendar
# Cal.com is REST API only — no SDK needed

# Observability
langfuse==4.5.0

# Scraping
playwright==1.58.0

# LLM
anthropic>=0.78.0     # Direct Anthropic SDK (replaces OpenRouter/openai)

# Evaluation
# tau2-bench — cloned separately at eval/tau2-bench/, not pip installed

# Data
# No pandas required — using stdlib csv for lightweight parsing
